requirements.txt
openpyxl==3.1.5
plotly==5.24.1
pandas==2.2.2
streamlit==1.37.1
numpy==1.26.4
Pillow==10.4.0
